+++ tests/subsys/storage/flash_map/CMakeLists.txt
@@ -4,5 +4,10 @@ cmake_minimum_required(VERSION 3.20.0)
 find_package(Zephyr REQUIRED HINTS $ENV{ZEPHYR_BASE})
 project(flash_map)
 
-FILE(GLOB app_sources src/*.c)
+if(NOT CONFIG_FLASH_AREA_CHECK_INTEGRITY_PSA AND NOT CONFIG_FLASH_AREA_CHECK_INTEGRITY_MBEDTLS)
+  FILE(GLOB app_sources src/main.c)
+else()
+  FILE(GLOB app_sources src/main_sha.c)
+endif()
+
 target_sources(app PRIVATE ${app_sources})
